backend/pyproject.toml
[tool.poetry]
name = "project-brief-backend"
version = "0.1.0"
description = "FastAPI service for Project Brief Agents"
authors = ["Usman Malik <malik_usman0185@hotmail.com>"]
readme = "backend/README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.121.1"
uvicorn = "^0.38.0"
motor = "^3.7.1"
pydantic-settings = "^2.11.0"
python-dotenv = "^1.2.1"
httpx = "^0.28.1"
langchain-community = "^0.4.1"
pypdf = "^6.2.0"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

